refactored code to show checkbox in case user is not confirmed
